add doc for streaks plotting + separate fct process_gradients
